requirements.txt
jira==3.4.1
python-dotenv==0.21.0
aiogram==2.22.2